schedule should be an abstract class
